Add Todo detail sheet and friendly date utility; enhance TodoController with update functionality
